+++ application/controllers/admin/participantsaction.php
@@ -2344,7 +2344,7 @@ $url .= "_view"; });
         $participantIdsString = Yii::app()->request->getPost('participant_id'); // TODO: This is a comma separated string of ids
         $participantIds = explode(",", $participantIdsString);
 
-        $surveyId = Yii::app()->request->getPost('surveyid');
+        $surveyId = (int) Yii::app()->request->getPost('surveyid');
 
         /**
          * mapped can take values like
